feat: Update API to only get exchange rates for KZT.
It was possible to get rates for any supported currency,
by calculating the rate through KZT.
For example USD to KZT = {usd_rate},
GBP to KZT = {gbp_rate}, so it was producing USD to GBP
rate by {usd_rate}/{gbp_rate}, which is correct math,
but Economy does not work that way!

So now such methods are removed, because wrong rates
were produced.